refactor(reducer): extract CommandPalette reduce-block
Move the 7 command-palette cases (toggleCommandPalette,
dismissCommandPalette, commandPaletteQueryChanged,
commandPaletteSelectIndex, commandPaletteSelectNext,
commandPaletteSelectPrevious, commandPaletteConfirm) out of the giant
AppReducer.body switch into a dedicated commandPaletteReducer block in
AppReducer+CommandPalette.swift, alongside the focus-handoff helper
(scheduleFocusAfterPaletteClose), the PaletteFocusID cancellation enum,
and the paletteFocusHandoffDelay constant -- all used only by these
cases.

This is a reduce-block over the shared AppReducer.State, NOT a child
reducer with its own State slice (the blueprint's documented fallback):
the handlers compute selection over state.commandPaletteItems (a
computed property over state.workspaces) and commandPaletteConfirm
writes state.activeWorkspaceID + lastAccessedAt directly. The three UI
fields (isCommandPaletteVisible, commandPaletteQuery,
commandPaletteSelectedIndex) and commandPaletteItems stay on
AppReducer.State.

Bodies, the direct activeWorkspaceID/lastAccessedAt writes + effect
ordering in confirm, and the .cancel/schedule focus-handoff logic are
moved verbatim. domain(of:) gains a .commandPalette arm and stays
exhaustive (no default). AppReducer.swift: 2466 -> 2360 lines. 1127
tests green.
